# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_verify_export_api_exports_verify_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_verify_export_api_exports_verify_post",
        "type": "object"
      },
      "ByteRange": {
        "additionalProperties": false,
        "properties": {
          "length": {
            "title": "Length",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          }
        },
        "required": [
          "offset",
          "length"
        ],
        "title": "ByteRange",
        "type": "object"
      },
      "Case": {
        "additionalProperties": false,
        "properties": {
          "case_number": {
            "title": "Case Number",
            "type": "string"
          },
          "created_utc": {
            "title": "Created Utc",
            "type": "string"
          },
          "fir_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fir Reference"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "lab": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lab"
          },
          "status": {
            "enum": [
              "open",
              "closed",
              "archived"
            ],
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_utc": {
            "title": "Updated Utc",
            "type": "string"
          }
        },
        "required": [
          "id",
          "case_number",
          "title",
          "fir_reference",
          "lab",
          "status",
          "created_utc",
          "updated_utc"
        ],
        "title": "Case",
        "type": "object"
      },
      "CaseCreate": {
        "additionalProperties": false,
        "properties": {
          "case_number": {
            "title": "Case Number",
            "type": "string"
          },
          "fir_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fir Reference"
          },
          "lab": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lab"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "case_number",
          "title"
        ],
        "title": "CaseCreate",
        "type": "object"
      },
      "CasePatch": {
        "additionalProperties": false,
        "properties": {
          "fir_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fir Reference"
          },
          "lab": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lab"
          },
          "status": {
            "anyOf": [
              {
                "enum": [
                  "open",
                  "closed",
                  "archived"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "CasePatch",
        "type": "object"
      },
...
      "ExportCreate": {
        "additionalProperties": false,
        "properties": {
          "channel": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel"
          },
          "from_norm_us": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Norm Us"
          },
          "recording_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recording Id"
          },
          "to_norm_us": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "To Norm Us"
          }
        },
        "title": "ExportCreate",
        "type": "object"
      },
      "ExportRecord": {
        "additionalProperties": false,
        "properties": {
          "case_id": {
            "title": "Case Id",
            "type": "string"
          },
          "channel": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel"
          },
          "created_utc": {
            "title": "Created Utc",
            "type": "string"
          },
          "examiner": {
            "title": "Examiner",
            "type": "string"
          },
          "file_path": {
            "title": "File Path",
            "type": "string"
          },
          "from_norm_us": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Norm Us"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "manifest_sha256": {
            "title": "Manifest Sha256",
            "type": "string"
          },
          "recording_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recording Id"
          },
          "signature_path": {
            "title": "Signature Path",
            "type": "string"
          },
          "to_norm_us": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "To Norm Us"
          }
        },
        "required": [
          "id",
          "case_id",
          "recording_id",
          "channel",
          "from_norm_us",
          "to_norm_us",
          "created_utc",
          "examiner",
          "file_path",
          "signature_path",
          "manifest_sha256"
        ],
        "title": "ExportRecord",
        "type": "object"
      },
      "ExportVerifyResult": {
        "additionalProperties": false,
        "properties": {
          "manifest": {
            "additionalProperties": true,
            "title": "Manifest",
            "type": "object"
          },
          "signature_valid": {
            "title": "Signature Valid",
            "type": "boolean"
          },
          "source_matches_registered_evidence": {
            "title": "Source Matches Registered Evidence",
            "type": "boolean"
          }
        },
        "required": [
          "signature_valid",
          "manifest",
          "source_matches_registered_evidence"
        ],
        "title": "ExportVerifyResult",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Job": {
        "additionalProperties": false,
        "properties": {
          "case_id": {
            "title": "Case Id",
            "type": "string"
          },
          "created_utc": {
            "title": "Created Utc",
            "type": "string"
          },
          "evidence_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Id"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "scan",
              "verify",
              "report",
              "export"
            ],
            "title": "Kind",
            "type": "string"
          },
          "log_lines": {
            "items": {
              "type": "string"
            },
            "title": "Log Lines",
            "type": "array"
          },
          "pct": {
            "title": "Pct",
            "type": "number"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/JobStage"
            },
            "title": "Stages",
            "type": "array"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "done",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "updated_utc": {
            "title": "Updated Utc",
            "type": "string"
          }
        },
        "required": [
          "id",
          "case_id",
          "evidence_id",
          "kind",
          "status",
          "stages",
          "pct",
          "created_utc",
          "updated_utc",
          "log_lines"
        ],
        "title": "Job",
        "type": "object"
      },
...
      "Recording": {
        "additionalProperties": false,
        "properties": {
          "byte_ranges": {
            "items": {
              "$ref": "#/components/schemas/ByteRange"
            },
            "title": "Byte Ranges",
            "type": "array"
          },
          "channel": {
            "title": "Channel",
            "type": "integer"
          },
          "deleted": {
            "title": "Deleted",
            "type": "boolean"
          },
          "end_ts_us": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Ts Us"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "image_id": {
            "title": "Image Id",
            "type": "string"
          },
          "source": {
            "enum": [
              "index",
              "carved",
              "inferred"
            ],
            "title": "Source",
            "type": "string"
          },
          "start_ts_us": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Ts Us"
          },
          "stream": {
            "enum": [
              "main",
              "sub"
            ],
            "title": "Stream",
            "type": "string"
          }
        },
        "required": [
          "id",
          "image_id",
          "channel",
          "stream",
          "start_ts_us",
          "end_ts_us",
          "byte_ranges",
          "source",
          "deleted"
        ],
        "title": "Recording",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Multi-vendor DVR/NVR forensic platform \u2014 API (docs/02-BACKEND.md).",
    "title": "Pramaan API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/cases": {
      "get": {
        "operationId": "list_cases_api_cases_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Case"
                  },
                  "title": "Response List Cases Api Cases Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Cases",
        "tags": [
          "cases"
        ]
      },
      "post": {
        "operationId": "create_case_api_cases_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaseCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Case"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Case",
        "tags": [
          "cases"
        ]
      }
    },
    "/api/cases/{cid}": {
      "get": {
        "operationId": "get_case_api_cases__cid__get",
        "parameters": [
          {
            "in": "path",
            "name": "cid",
            "required": true,
            "schema": {
              "title": "Cid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Case"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Case",
        "tags": [
          "cases"
        ]
      },
      "patch": {
        "operationId": "patch_case_api_cases__cid__patch",
        "parameters": [
          {
            "in": "path",
            "name": "cid",
            "required": true,
            "schema": {
              "title": "Cid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CasePatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Case"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Case",
        "tags": [
          "cases"
        ]
      }
    },
...
    "/api/cases/{cid}/exports": {
      "post": {
        "operationId": "create_export_api_cases__cid__exports_post",
        "parameters": [
          {
            "in": "path",
            "name": "cid",
            "required": true,
            "schema": {
              "title": "Cid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Export",
        "tags": [
          "exports"
        ]
      }
    },
...
    "/api/cases/{cid}/jobs": {
      "get": {
        "operationId": "list_jobs_api_cases__cid__jobs_get",
        "parameters": [
          {
            "in": "path",
            "name": "cid",
            "required": true,
            "schema": {
              "title": "Cid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Job"
                  },
                  "title": "Response List Jobs Api Cases  Cid  Jobs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Jobs",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/api/cases/{cid}/recordings": {
      "get": {
        "operationId": "list_recordings_api_cases__cid__recordings_get",
        "parameters": [
          {
            "in": "path",
            "name": "cid",
            "required": true,
            "schema": {
              "title": "Cid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "channel",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Channel"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "index",
                    "carved",
                    "inferred"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "in": "query",
            "name": "deleted",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Deleted"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Recording"
                  },
                  "title": "Response List Recordings Api Cases  Cid  Recordings Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Recordings",
        "tags": [
          "recordings"
        ]
      }
    },
...
    "/api/clips/{clip_id}/stream": {
      "get": {
        "operationId": "stream_clip_api_clips__clip_id__stream_get",
        "parameters": [
          {
            "in": "path",
            "name": "clip_id",
            "required": true,
            "schema": {
              "title": "Clip Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Clip",
        "tags": [
          "clips"
        ]
      }
    },
...
    "/api/exports/verify": {
      "post": {
        "operationId": "verify_export_api_exports_verify_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_verify_export_api_exports_verify_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportVerifyResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Export",
        "tags": [
          "exports"
        ]
      }
    },
    "/api/exports/{xid}/file": {
      "get": {
        "operationId": "export_file_api_exports__xid__file_get",
        "parameters": [
          {
            "in": "path",
            "name": "xid",
            "required": true,
            "schema": {
              "title": "Xid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export File",
        "tags": [
          "exports"
        ]
      }
    },
...
    "/api/jobs/{jid}": {
      "get": {
        "operationId": "get_job_api_jobs__jid__get",
        "parameters": [
          {
            "in": "path",
            "name": "jid",
            "required": true,
            "schema": {
              "title": "Jid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/api/recordings/{rid}": {
      "get": {
        "operationId": "get_recording_api_recordings__rid__get",
        "parameters": [
          {
            "in": "path",
            "name": "rid",
            "required": true,
            "schema": {
              "title": "Rid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Recording"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Recording",
        "tags": [
          "recordings"
        ]
      }
    },
...
  }
}